+++ mimesis/providers/generic.py
@@ -1,10 +1,10 @@
 """Provides all at one."""
 
 import inspect
-from typing import Any, Callable, List
+from typing import Any, List, Sequence
 
 from mimesis.providers.address import Address
-from mimesis.providers.base import BaseDataProvider
+from mimesis.providers.base import BaseDataProvider, BaseProvider
 from mimesis.providers.business import Business
 from mimesis.providers.choice import Choice
 from mimesis.providers.clothing import ClothingSize
@@ -26,7 +26,6 @@ from mimesis.providers.structure import Structure
 from mimesis.providers.text import Text
 from mimesis.providers.transport import Transport
 from mimesis.providers.units import UnitSystem
-from mimesis.typing import Seed
 
 __all__ = ['Generic']
 
@@ -104,17 +103,21 @@ class Generic(BaseDataProvider):
         :raises TypeError: if cls is not class.
         """
         if inspect.isclass(cls):
+            if not issubclass(cls, BaseProvider):
+                raise TypeError('The provider must inherit BaseProvider.')
+
             name = ''
             if hasattr(cls, 'Meta'):
                 if inspect.isclass(cls.Meta) and hasattr(cls.Meta, 'name'):
                     name = cls.Meta.name
             else:
                 name = cls.__name__.lower()
+
             setattr(self, name, cls())
         else:
             raise TypeError('Provider must be a class')
 
-    def add_providers(self, *providers) -> None:
+    def add_providers(self, *providers: Sequence[BaseProvider]) -> None:
         """Add a lot of custom providers to Generic() object.
 
         :param providers: Custom providers.